Access controll users. Relationship between Users and Todos

+++ app/Controllers/Http/TodoController.js
@@ -20,11 +20,16 @@ class TodoController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async index ({ request, response, view }) {
-    const todos = await Todo.all()
+  async index ({ request, auth, response, view }) {
+    //const todos = await Todo.all()
+    const todos = await Todo
+      .query()
+      .where('user_id', auth.user.id)
+      .fetch()
 
     return view.render('index', {
-      todos: todos.toJSON()
+      todos: todos.toJSON(),
+      name: auth.user.username
     })
   }
 
@@ -48,30 +53,9 @@ class TodoController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async store ({ request, session, response }) {
-
-    // const rules = {
-    //   addTodo: 'required|min:3'
-    // }
-    //
-    // const messages = {
-    //   'addTodo.required': 'The add todo field is required',
-    //   'addTodo.min': 'The add todo field requires at least 3 characters'
-    // }
-    //
-    // const validation = await validate(request.all(), rules, messages)
-    //
-    // if (validation.fails()) {
-    //   session
-    //     .withErrors(validation.messages())
-    //     .flashAll()
-    //
-    //   return response.redirect('back')
-    // }
-
-
-
+  async store ({ request, auth, session, response }) {
     const todo = await Todo.create({
+      user_id: auth.user.id,
       title: request.input('addTodo')
      })
      session.flash({ successMessage: 'Todo was added!' })
@@ -99,9 +83,12 @@ class TodoController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async edit ({ params, request, response, view }) {
+  async edit ({ params, auth, request, response, view }) {
     const todo = await Todo.findBy('id', params.id)
-    //console.log(todo);
+
+    if(auth.user.id !== todo.user_id) {
+      return 'You do not have permission to do this'
+    }
 
     return view.render('edit', {
       todo
@@ -116,8 +103,13 @@ class TodoController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async update ({ params, session, request, response }) {
+  async update ({ params, auth, session, request, response }) {
     const todo = await Todo.findOrFail(params.id)
+
+    if(auth.user.id !== todo.user_id) {
+      return 'You do not have permission to do this'
+    }
+
     todo.title = request.input('editTodo')
     todo.completed = request.input('completedCheck') === 'on' ? true : false
     await todo.save()
@@ -134,9 +126,13 @@ class TodoController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async destroy ({ params, session, request, response }) {
+  async destroy ({ params, auth, session, request, response }) {
 
     const todo = await Todo.findOrFail(params.id)
+
+    if(auth.user.id !== todo.user_id) {
+      return 'You do not have permission to do this'
+    }
 
     await todo.delete()
